Split module header in Interface into multiple files
